VR: Support for new MouseUDP_1_ev.py

+++ Assets/Scripts/Game/SphereInput.cs
@@ -3,17 +3,17 @@ using System.Collections;
 
 public class SphereInput {
 
-    public sbyte mouse1X;
-    public sbyte mouse1Y;
-    public sbyte mouse2X;
-    public sbyte mouse2Y;
+	public int mouse1X;
+	public int mouse1Y;
+	public int mouse2X;
+	public int mouse2Y;
 
     public SphereInput()
     {
 
     }
 
-    public SphereInput(sbyte _m1x, sbyte _m1y, sbyte _m2x, sbyte _m2y)
+	public SphereInput(int _m1x, int _m1y, int _m2x, int _m2y)
     {
         this.mouse1X = _m1x;
         this.mouse1Y = _m1y;
